+++ mithril-signer/src/configuration.rs
@@ -116,6 +116,9 @@ pub struct Configuration {
 
     /// The maximum number of roll forwards during a poll of the block streamer when importing transactions.
     pub cardano_transactions_block_streamer_max_roll_forwards_per_poll: usize,
+
+    /// Preloading refresh interval in seconds
+    pub preloading_refresh_interval_in_seconds: u64,
 }
 
 impl Configuration {
@@ -153,6 +156,7 @@ impl Configuration {
             enable_transaction_pruning: false,
             transactions_import_block_chunk_size: BlockNumber(1000),
             cardano_transactions_block_streamer_max_roll_forwards_per_poll: 1000,
+            preloading_refresh_interval_in_seconds: 60,
         }
     }